==> scripts/Crossdate Workflow/4_Build_Shiny_App.R <==
##----------------------------------------------------------------------------------------------------------------------------------#
#Script 4 of 4 
#Script by: Jocelyn Garcia
#Project: Living Collections Wood Archive 
#Purpose:Create a web application housing the master chronology from "2.5_Build_chronology.R."
#         Web app should allow for isolation of certain time period inputed by the user.
#Helpful resource: https://mastering-shiny.org/basic-app.html


#install.packages("shiny") #might need to download if first time

library(shiny)
